requirements.txt
pymongo==4.0.1
python-dotenv==0.19.2
apscheduler==3.8.1
tweepy==4.4.0
pydantic~=1.8.2
fastapi==0.70.0
uvicorn==0.15.0
